% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{clean_matches_data}
\alias{clean_matches_data}
\title{Clean the matches data from deuce atp_maches and wta_matches}
\usage{
clean_matches_data(atp_matches = NULL, wta_matches = NULL, start_year,
  end_year, include_quals = FALSE, level = "Grand Slams")
}
\arguments{
\item{atp_matches}{atp_matches object from R package deuce (Men's).  Default is NULL}

\item{wta_matches}{atp_matches object from R package deuce (Women's).  Default is NULL}

\item{start_year}{start year (earliest is 1968)}

\item{end_year}{end year (latest available is 2017)}

\item{include_quals}{logical to include qualifying rounds.  Default is FALSE}

\item{level}{level of tournament, corresponding to column from atp_matches/wta_matches.  Default is "Grand Slams"}
}
\value{
cleaned and subsetted data set, combined if there is both an atp_matches and wta_data set.  A new column called "league" is added and has value "ATP" or "WTA".  We also clean the tourney_name making a new column called tournament
}
\description{
Clean the matches data from deuce atp_maches and wta_matches
}
