% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{combine_fields}
\alias{combine_fields}
\title{Combine fields to have easier access}
\usage{
combine_fields(gs_join)
}
\arguments{
\item{gs_joint}{result from join_gs_pbp}
}
\value{
combined fields
}
\description{
Combine fields to have easier access
}
