% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{extract_sgp}
\alias{extract_sgp}
\title{Extract set, games, and points for winner and loser}
\usage{
extract_sgp(matches)
}
\arguments{
\item{matches}{object from deuce package or clean_matches_data.}
}
\value{
matches with extra columns w_pointswon, l_pointswon, w_gameswon, l_gameswon, w_setswon, l_setswon
}
\description{
Extract set, games, and points for winner and loser
}
