% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{join_gs_pbp}
\alias{join_gs_pbp}
\title{Join gs and gs_pbp together}
\usage{
join_gs_pbp(gs_pbp, gs)
}
\arguments{
\item{gs_pbp}{point by point summarized by matches}

\item{gs}{subsetted gs object}
}
\value{
joined data frame by tournament, year, and players, and added a week column
}
\description{
Join gs and gs_pbp together
}
