% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{matches_to_player_data}
\alias{matches_to_player_data}
\title{Turn the matches data into player data for modelling}
\usage{
matches_to_player_data(matches, w_opp_vars = "loser_rank",
  l_opp_vars = "winner_rank", opp_var_names = "opponent_rank")
}
\arguments{
\item{matches}{object from deuce package or clean_matches_data.}

\item{w_opp_vars}{opponent variable base name for the winner to keep.  Default is "loser_rank"}

\item{l_opp_vars}{opponent variable base name for the loser to keep.  Default is "winner_rank"}

\item{opp_var_names}{what to rename the opponent variables}
}
\value{
formatted matches so each row refers to one player, their attributes, and their opponents variables along with the other tournament variables
}
\description{
Turn the matches data into player data for modelling
}
\details{
All variables to be removed must start with "w_" "W" or "winner_" or "l_", "L", or "loser_"
}
