% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-individs.R
\name{model_individual}
\alias{model_individual}
\title{Model the individuals using stepwise linear regression}
\usage{
model_individual(player_name = "Roger Federer",
  data = gs_partial_players, ref = "Wimbledon", start_lower = FALSE,
  test_prop = 0, min_year = 2013, max_year = 2017, seed = NULL)
}
\arguments{
\item{player_name}{player name}

\item{data}{default is gs_partial_players}

\item{ref}{reference court.  Default is French Open}

\item{start_lower}{logical indicating whether the base model should be the lower model.}
}
\value{
list step model output, the final model on the training set,
the subsetted data, the training indices, and the number of total observations in the subsetted data (n_obs), the number of training obs (n_train), and number of testing obs (n_test)
}
\description{
Model the individuals using stepwise linear regression
}
