% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{summarize_pbp}
\alias{summarize_pbp}
\title{Summarize the point by point data}
\usage{
summarize_pbp(pbp, start_year = 2013, end_year = 2017)
}
\arguments{
\item{pbp}{point by point data from deuce package}
}
\value{
summary of point by point data in terms of unique matches
}
\description{
Summarize the point by point data
}
